requirements.txt
flask
requests
beautifulsoup4
python-dotenv
python-logging-loki
gunicorn
